% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm.sampling.R
\name{comm.sampling}
\alias{comm.sampling}
\title{Converting simulation results into site-by-species matrix}
\usage{
comm.sampling(x, type = c("full", "random"), size)
}
\arguments{
\item{x}{community and trait data matrix produced by \code{\link{comm.simul}}
function}

\item{type}{Type of sampling. If \code{type=="full"} sample size equals to community size.
It simply converts \code{x} into a site-by-species matrix
If \code{type=="random"}, it applies random sampling by calling (\code{\link[vegan]{rrarefy}}) function}

\item{size}{Number of individuals in the random samples. It should be smaller
than number of individuals in simulated (sub)communities. Otherwise,
\code{x} is converted into a site-by-species matrix without (re)sampling}
}
\value{
A site-by-species matrix containing abundances.
}
\description{
Converts simulation result into site-by-species matrix of abundances,
and optionally in the same step simulates random sampling with fixed number of individuals.
}
\details{
If \code{type=="full"} it simply converts simulation results from long to wide format.
If \code{type=="random"} it randomly selects \code{size} individuals in each (sub)community
and abundances in these samples are converted into site-by-species matrix format.
}
\examples{
x<-comm.simul(S=20, J=30)
str(x$final.community)

w<-comm.sampling(x$final.community,type="full")
str(w)

w.rarefied<-comm.sampling(x$final.community,type="random",size=10)
rowSums(w)
rowSums(w.rarefied)

}
